Check if Settle is being run as root (#19)
If Settle is being run as root, then initialize the package manager with 
sudo == False so the commands are run directly without asking for root
permissions. Pass the sudo argument of package manager child classes to the
parent constructor through super, so we don't need to define the sudo attribute
on each child class.