Enable list-packages functionality
Add Asker the ability to show question for categories or packages if
list-packages is False or True, respectively. Add Asker a method for
creating the list of packages that must be installed. Rename
default_packages to default_categories: it's a better name for a list
that contains default categories and not package names.